[refactor] Update imports and Gradle files after `core:ktor` module relocation
