+++ data/market-remote/src/main/java/ir/composenews/remotedatasource/api/MarketsApi.kt
@@ -2,7 +2,7 @@
 
 package ir.composenews.remotedatasource.api
 
-import ir.composenews.ktor.ApiResponse
+import ir.composenews.network.ApiResponse
 import ir.composenews.remotedatasource.dto.MarketChartResponse
 import ir.composenews.remotedatasource.dto.MarketDetailResponse
 import ir.composenews.remotedatasource.dto.MarketResponse
