+++ data/market-repository/src/main/java/ir/composenews/data/repository/MarketRepositoryImpl.kt
@@ -12,17 +12,17 @@ import ir.composenews.domain.model.Chart
 import ir.composenews.domain.model.Market
 import ir.composenews.domain.model.MarketDetail
 import ir.composenews.domain.repository.MarketRepository
-import ir.composenews.ktor.Errors
-import ir.composenews.ktor.Resource
-import ir.composenews.ktor.mapMessageStatusCode
-import ir.composenews.ktor.onError
-import ir.composenews.ktor.onException
-import ir.composenews.ktor.statusCode
-import ir.composenews.ktor.suspendMap
-import ir.composenews.ktor.suspendOnError
-import ir.composenews.ktor.suspendOnException
-import ir.composenews.ktor.suspendOnSuccess
 import ir.composenews.localdatasource.database.MarketDao
+import ir.composenews.network.Errors
+import ir.composenews.network.Resource
+import ir.composenews.network.mapMessageStatusCode
+import ir.composenews.network.onError
+import ir.composenews.network.onException
+import ir.composenews.network.statusCode
+import ir.composenews.network.suspendMap
+import ir.composenews.network.suspendOnError
+import ir.composenews.network.suspendOnException
+import ir.composenews.network.suspendOnSuccess
 import ir.composenews.remotedatasource.api.MarketsApi
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
