+++ domain/market/src/main/java/ir/composenews/domain/repository/MarketRepository.kt
@@ -3,8 +3,8 @@ package ir.composenews.domain.repository
 import ir.composenews.domain.model.Chart
 import ir.composenews.domain.model.Market
 import ir.composenews.domain.model.MarketDetail
-import ir.composenews.ktor.Errors
-import ir.composenews.ktor.Resource
+import ir.composenews.network.Errors
+import ir.composenews.network.Resource
 import kotlinx.coroutines.flow.Flow
 
 interface MarketRepository {
