+++ domain/market/src/main/java/ir/composenews/domain/use_case/GetMarketChartUseCase.kt
@@ -4,8 +4,8 @@ package ir.composenews.domain.use_case
 
 import ir.composenews.domain.model.Chart
 import ir.composenews.domain.repository.MarketRepository
-import ir.composenews.ktor.Errors
-import ir.composenews.ktor.Resource
+import ir.composenews.network.Errors
+import ir.composenews.network.Resource
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
